* Fixed png bug.  * Added height to terrain.  * Added new terrain only map.  * Finished all terrain stuff.  * Removed unused options.

+++ tiles.h
@@ -85,7 +85,7 @@ int get_tile_type(int x, int y);
  * \param y						Y coordinate
  * \callgraph
  */
-float get_tile_height(const float x, const float y);
+float get_tile_height(const int x, const int y);
 
 /*!
  * \ingroup 	tile
